api/requirements.txt
Django==4.2.5
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
firebase-admin==6.2.0
uvicorn==0.25.0
python-telegram-bot==20.7
urllib3==1.26.16
whitenoise==6.5.0
adrf==0.1.3
yookassa==3.0.0
aiohttp==3.9.1
